# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Wishmill API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/_wishmill/v1",
    "paths": {
        "/auth/checkToken": {
            "post": {
                "summary": "Check token validity",
                "parameters": [
                    {
                        "description": "Authorize data from oidc provider",
                        "name": "Session",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Session"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Errormsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Errormsg"
                        }
                    }
                }
            }
        },
        "/auth/obtainToken": {
            "post": {
                "summary": "Generate a session token",
                "parameters": [
                    {
                        "description": "Authorize data from oidc provider",
                        "name": "Login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.LoginBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Session"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Errormsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Errormsg"
                        }
                    }
                }
            }
        },
        "/auth/oidc_providers": {
            "get": {
                "summary": "Get oidc authentication providers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/config.Oidc_provider"
                            }
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "summary": "Get app health",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Errormsg": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "api.LoginBody": {
            "type": "object",
            "required": [
                "code",
                "provider",
                "redirect_url"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "redirect_url": {
                    "type": "string"
                }
            }
        },
        "api.Session": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "api.User": {
            "type": "object",
            "properties": {
                "authProvider": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sub": {
                    "type": "string"
                }
            }
        },
        "config.Oidc_provider": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}